added pin functionality no toggle
