+++ ui/src/ContainerHealth.tsx
@@ -1,5 +1,4 @@
-import react from 'react';
-import Container from './Container';
+import React from 'react';
 
 export default function ContainerHealth() {
   return (
